Restrict the meta package version to address fluter sdk limitations
Currently using this package in flutter results in a version incompatability owing to restrictions on the meta package from flutter_test:

Because every version of flutter_test from sdk depends on meta 1.1.6 and every version of vin_decoder depends on meta ^1.1.7, flutter_test from sdk is incompatible with vin_decoder.

We temporarily constrain the version to 1.1.6 until this is resolved.